requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
curlify==2.2.1
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
loguru==0.7.0
minio==7.1.15
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pika==1.3.2
pluggy==1.0.0
psycopg2-binary==2.9.6
pycparser==2.21
pysocks==1.7.1
pytest==7.3.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-interface==1.6.1
requests-oauthlib==1.3.1
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
testit-importer-allure==1.3.3
tomli==2.0.1
tqdm==4.65.0
trio-websocket==0.10.3
trio==0.22.0
urllib3==2.0.3
urllib3[socks]==2.0.3
webdriver-manager==3.8.6
win32-setctime==1.1.0
wsproto==1.2.0
xmltodict==0.13.0
